+++ src/api/resources/organizations/resources/placements/client/requests/index.ts
@@ -1,1 +1,2 @@
+export type { GetPlacementRequest } from "./GetPlacementRequest.js";
 export type { ListPlacementsRequest } from "./ListPlacementsRequest.js";